Move gallery logic to vanilla JS module
