+++ resources/js/app.js
@@ -1,4 +1,5 @@
 import './bootstrap';
 
 
+import "./gallery";
 // Custom scripts can be added here
